@@ -1,45 +1,49 @@
 {
-    "name": "dashboard",
-    "version": "1.0.0",
-    "description": "",
-    "main": "./src/index.js",
-    "scripts": {
-      "dev": "webpack serve --mode development --config config/webpack.config.js",
-      "build": "webpack --mode production --config config/webpack.config.js",
-      "start": "webpack serve --open --mode development --config config/webpack.config.js",
-      "test": "jest"
+  "name": "dashboard",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "jest",
+    "build": "webpack build --config ./config/webpack.config.js",
+    "start": "webpack-dev-server --mode development --config config/webpack.config.js",
+    "test-watch": "jest --watch"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@babel/core": "^7.18.6",
+    "@babel/plugin-transform-runtime": "^7.18.6",
+    "@babel/preset-env": "^7.18.6",
+    "@babel/preset-react": "^7.18.6",
+    "babel-loader": "^8.2.5",
+    "css-loader": "^6.7.1",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.5.0",
+    "image-webpack-loader": "^8.1.0",
+    "jest": "^28.1.3",
+    "style-loader": "^3.3.1",
+    "webpack": "^5.73.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.9.3"
+  },
+  "dependencies": {
+    "@babel/runtime": "^7.18.6",
+    "@zarconontol/enzyme-adapter-react-18": "^0.7.3",
+    "inline-source-map": "^0.6.2",
+    "prop-types": "^15.8.1"
+  },
+  "jest": {
+    "setupFiles": [
+      "<rootDir>/config/setupTests.js"
+    ],
+    "moduleNameMapper": {
+      "\\.(jpg|ico|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
+      "\\.(css|less)$": "<rootDir>/__mocks__/styleMock.js"
     },
-    "keywords": [],
-    "author": "",
-    "license": "ISC",
-    "devDependencies": {
-      "@babel/core": "^7.12.16",
-      "@babel/preset-env": "^7.12.16",
-      "@babel/preset-react": "^7.12.13",
-      "babel-loader": "^8.2.2",
-      "css-loader": "^5.0.2",
-      "enzyme": "^3.11.0",
-      "enzyme-adapter-react-16": "^1.15.6",
-      "file-loader": "^6.2.0",
-      "image-webpack-loader": "^7.0.1",
-      "jest": "^26.6.3",
-      "style-loader": "^2.0.0",
-      "webpack": "^5.22.0",
-      "webpack-cli": "^4.5.0",
-      "webpack-dev-server": "^3.11.2"
-    },
-    "dependencies": {
-      "prop-types": "^15.7.2",
-      "react": "^17.0.1",
-      "react-dom": "^17.0.1"
-    },
-    "jest": {
-      "setupFiles": [
-        "<rootDir>/config/setupTests.js"
-      ],
-      "moduleNameMapper": {
-        "\\.(css)$": "<rootDir>/__mocks__/styleMock.js",
-        "\\.(jpg|png)$": "<rootDir>/__mocks__/fileMock.js"
-      }
+    "transform": {
+      "\\.js$": "<rootDir>/node_modules/babel-jest"
     }
   }
+}
