@@ -1,40 +1,26 @@
-import React from 'react';
-import './Notifications.css';
-import close_icon from '../assets/close-icon.png';
-import { getLatestNotification } from '../utils/utils';
+import React from "react";
+import "./Notifications.css";
+import closeIcon from "../assets/close-icon.png";
+import { getLatestNotification } from "../utils/utils";
 
-const Notifications = () => {
+function Notifications() {
   return (
-    <div className='Notifications'>
+    <div className="Notifications">
+      <button
+        style={{ color: "#3a3a3a", fontWeight: "bold", background: "none", border: "none", fontSize: "10px", position: "absolute", right: "2px", top: "2px", cursor: "pointer" }}
+        aria-label="Close"
+        onClick={console.log("Close button has been clicked")}
+      >
+        <img src={closeIcon} alt="closeIcon" width="10px" />
+      </button>
       <p>Here is the list of notifications</p>
       <ul>
-        <li data-priority='default'>New course available</li>
-        <li data-priority='urgent'>New resume available</li>
-        <li data-priority='urgent'>
-          <div
-            dangerouslySetInnerHTML={{ __html: `${getLatestNotification()}` }}
-          ></div>
-        </li>
+        <li data="default">New course available</li>
+        <li data="urgent">New resume available</li>
+        <li data="urgent" dangerouslySetInnerHTML={{ __html: getLatestNotification() }}></li>
       </ul>
-      <button
-        type='button'
-        aria-label='Close'
-        onClick={() => console.log('Close button has been clicked')}
-        style={{
-          display: 'inline-block',
-          position: 'absolute',
-          top: '16px',
-          right: '16px',
-          background: 0,
-          border: 0,
-          outline: 'none',
-          cursor: 'pointer',
-        }}
-      >
-        <img src={close_icon} alt='' style={{ width: '8px', height: '8px' }} />
-      </button>
     </div>
   );
-};
+}
 
 export default Notifications;
