@@ -1,4 +1,4 @@
-import Enzyme from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
+import { configure } from "enzyme";
+import Adapter from "@zarconontol/enzyme-adapter-react-18";
 
-Enzyme.configure({ adapter: new Adapter() });
+configure({ adapter: new Adapter() });
