@@ -1,13 +1,15 @@
-import React from 'react';
-import './Header.css';
-import logo from '../assets/holberton-logo.jpg';
+import React from "react";
+import logo from "../assets/holberton-logo.jpg";
+import "./Header.css";
 
 function Header() {
   return (
-    <header className='header'>
-      <img src={logo} alt='logo' />
-      <h1>School dashboard</h1>
-    </header>
+    <>
+      <div className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <h1>School dashboard</h1>
+      </div>
+    </>
   );
 }
 
