@@ -1,16 +1,20 @@
-import React from 'react';
-import './Login.css';
+import React from "react";
+import "./Login.css";
 
 function Login() {
   return (
-    <main role='main' className='login'>
-      <p>Login to access the full dashboard</p>
-      <label htmlFor='email'>Email</label>
-      <input type='email' name='email' id='email' />
-      <label htmlFor='password'>Password</label>
-      <input type='password' name='password' id='password' />
-      <button type='button'>OK</button>
-    </main>
+    <React.Fragment>
+      <div className="App-body">
+        <p>Login to access the full dashboard</p>
+        <form>
+          <label htmlFor="email">Email:</label>
+          <input type="email" name="email"></input>
+          <label htmlFor="password">Password:</label>
+          <input type="password" name="password"></input>
+          <button>OK</button>
+        </form>
+      </div>
+    </React.Fragment>
   );
 }
 
