@@ -1,17 +1,29 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import Notifications from './Notifications';
+import React from "react";
+import { shallow } from "enzyme";
+import Notifications from "./Notifications";
 
-const wrapper = shallow(<Notifications />);
+describe("Notification component tests", () => {
+  it("renders Notification component without crashing", () => {
+    const notification = shallow(<Notifications />);
 
-it('renders without crashing', () => {
-  shallow(<Notifications />);
-});
-it('renders three list items', () => {
-  expect(wrapper.find('li').children().length).toEqual(3);
-});
-it('renders the <p>', () => {
-  expect(
-    wrapper.containsMatchingElement(<p>Here is the list of notifications</p>)
-  ).toBeTruthy();
+    expect(notification).toBeDefined();
+  });
+
+  it("renders ul", () => {
+    const notification = shallow(<Notifications />);
+
+    expect(notification.find("ul")).toBeDefined();
+  });
+
+  it("renders three list items", () => {
+    const notification = shallow(<Notifications />);
+
+    expect(notification.find("li")).toHaveLength(3);
+  });
+
+  it("renders correct text", () => {
+    const notification = shallow(<Notifications />);
+
+    expect(notification.find("p").text()).toBe("Here is the list of notifications");
+  });
 });
