@@ -1,30 +1,13 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React from "react";
+import "./Notifications.css";
 
 function NotificationItem({ type, html, value }) {
-  let li;
-
-  value
-    ? (li = <li data-notification-type={type}>{value}</li>)
-    : (li = (
-        <li data-notification-type={type} dangerouslySetInnerHTML={html}></li>
-      ));
-
-  return li;
+  return (
+    <>
+      {type && value ? <li data-notification-type={type}>{value}</li> : null}
+      {html ? <li data-urgent dangerouslySetInnerHTML={{ __html: html }}></li> : null}
+    </>
+  );
 }
-
-NotificationItem.defaultProps = {
-  type: 'default',
-  value: '',
-  html: {},
-};
-
-NotificationItem.propTypes = {
-  type: PropTypes.string,
-  value: PropTypes.string,
-  html: PropTypes.shape({
-    __html: PropTypes.string,
-  }),
-};
 
 export default NotificationItem;
