@@ -1,37 +1,21 @@
-import React from 'react';
-import './CourseList.css';
-import CourseListRow from './CourseListRow';
+import React from "react";
+import "./CourseList.css";
+import CourseListRow from "./CourseListRow";
 
-const CourseList = () => {
+function CourseList() {
   return (
-    <table className='course-list'>
+    <table id="CourseList">
       <thead>
-        <CourseListRow textFirstCell='Available courses' isHeader={true} />
-        <CourseListRow
-          textFirstCell='Course name'
-          textSecondCell='Credit'
-          isHeader={true}
-        />
+        <CourseListRow textFirstCell="Available courses" textSecondCell={null} isHeader={true} />
+        <CourseListRow textFirstCell="Course name" textSecondCell="Credit" isHeader={true} />
       </thead>
       <tbody>
-        <CourseListRow
-          textFirstCell='ES6'
-          textSecondCell='60'
-          isHeader={false}
-        />
-        <CourseListRow
-          textFirstCell='Webpack'
-          textSecondCell='20'
-          isHeader={false}
-        />
-        <CourseListRow
-          textFirstCell='React'
-          textSecondCell='40'
-          isHeader={false}
-        />
+        <CourseListRow textFirstCell="ES6" textSecondCell="60" isHeader={false} />
+        <CourseListRow textFirstCell="Webpack" textSecondCell="20" isHeader={false} />
+        <CourseListRow textFirstCell="React" textSecondCell="40" isHeader={false} />
       </tbody>
     </table>
   );
-};
+}
 
 export default CourseList;
